# doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "RAAR Radio Archive API",
    "description": "RAAR Radio Archive API. Some endpoints are public, other are restricted to admins.",
    "license": {
      "name": "AGPL"
    }
  },
  "consumes": [
    "application/vnd.api+json"
  ],
  "produces": [
    "application/vnd.api+json"
  ],
  "securityDefinitions": {
    "http_token": {
      "type": "basic",
      "description": "API token is passed as HTTP token authentication header: `Authorization: Token token=\"abc\"`"
    },
    "api_token": {
      "type": "apiKey",
      "name": "api_token",
      "in": "query",
      "description": "API token is passed as a query parameter"
    }
  },
...
  "parameters": {
    "page_number": {
      "name": "page[number]",
      "in": "query",
      "description": "The page number of the list.",
      "required": false,
      "type": "integer"
    },
    "page_size": {
      "name": "page[size]",
      "in": "query",
      "description": "Maximum number of entries that are returned per page. Defaults to 50, maximum is 500.",
      "required": false,
      "type": "integer"
    },
    "sort": {
      "name": "sort",
      "in": "query",
      "description": "Name of the sort field, optionally prefixed with a `-` for descending order.",
      "required": false,
      "type": "string"
    },
    "q": {
      "name": "q",
      "in": "query",
      "description": "Query string to search for.",
      "required": false,
      "type": "string"
    }
  },
  "paths": {
...
    "/broadcasts": {
      "get": {
        "description": "Searches and returns a list of broadcasts.",
        "tags": [
          "broadcast",
          "public"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/q"
          },
          {
            "$ref": "#/parameters/page_number"
          },
          {
            "$ref": "#/parameters/page_size"
          },
          {
            "$ref": "#/parameters/sort"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Broadcast"
                  }
                },
                "included": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Show"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/broadcasts/{year}/{month}/{day}/{hour}{minute}{second}": {
      "get": {
        "description": "Returns a list of broadcasts at the given date/time span.",
        "tags": [
          "broadcast",
          "public"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "The four-digit year to get the broadcasts for.",
            "required": true,
            "type": "integer"
          },
          {
            "name": "month",
            "in": "path",
            "description": "Optional two-digit month to get the broadcasts for. Requires all preceeding parameters.",
            "required": true,
            "type": "integer"
          },
          {
            "name": "day",
            "in": "path",
            "description": "Optional two-digit day to get the broadcasts for. Requires all preceeding parameters.",
            "required": true,
            "type": "integer"
          },
          {
            "name": "hour",
            "in": "path",
            "description": "Optional two-digit hour to get the broadcasts for. Requires all preceeding parameters.",
            "required": true,
            "type": "integer"
          },
          {
            "name": "minute",
            "in": "path",
            "description": "Optional two-digit minute to get the broadcasts for. Requires all preceeding parameters.",
            "required": true,
            "type": "integer"
          },
          {
            "name": "second",
            "in": "path",
            "description": "Optional two-digit second to get the broadcasts for. Requires all preceeding parameters.",
            "required": true,
            "type": "integer"
          },
          {
            "$ref": "#/parameters/q"
          },
          {
            "$ref": "#/parameters/page_number"
          },
          {
            "$ref": "#/parameters/page_size"
          },
          {
            "$ref": "#/parameters/sort"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Broadcast"
                  }
                },
                "included": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Show"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/shows/{show_id}/broadcasts": {
      "get": {
        "description": "Returns a list of broadcasts of the given show.",
        "tags": [
          "broadcast",
          "public"
        ],
        "parameters": [
          {
            "name": "show_id",
            "in": "path",
            "description": "ID of the show to list the broadcasts for",
            "required": true,
            "type": "integer"
          },
          {
            "$ref": "#/parameters/q"
          },
          {
            "$ref": "#/parameters/page_number"
          },
          {
            "$ref": "#/parameters/page_size"
          },
          {
            "$ref": "#/parameters/sort"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Broadcast"
                  }
                },
                "included": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Show"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "get": {
        "description": "Get the user object of the currently logged in user.",
        "tags": [
          "user"
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "properties": {
                "data": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "401": {
            "description": "not authorized"
          }
        }
      },
      "post": {
        "description": "Login with username and password. Returns the user object including the api_token for further requests.",
        "tags": [
          "user"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "description": "The username of the user to login.",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "The password of the user to login.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "properties": {
                "data": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "401": {
            "description": "not authorized"
          }
        }
      }
    },
    "/login/api_key": {
      "put": {
        "description": "Regenerates the api key of the current user.",
        "tags": [
          "user"
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "properties": {
                "data": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "401": {
            "description": "not authorized"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Broadcast": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "attributes": {
          "properties": {
            "label": {
              "type": "string"
            },
            "started_at": {
              "type": "string",
              "format": "date-time"
            },
            "finished_at": {
              "type": "string",
              "format": "date-time"
            },
            "people": {
              "type": "string"
            },
            "details": {
              "type": "string"
            }
          }
        },
        "relationships": {
          "properties": {
            "show": {
              "properties": {
                "data": {
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "type": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "Show": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "attributes": {
          "properties": {
            "name": {
              "type": "string"
            },
            "details": {
              "type": "string"
            }
          }
        },
        "links": {
          "properties": {
            "self": {
              "type": "string",
              "format": "url",
              "readOnly": true
            }
          }
        }
      }
    },
...
    "User": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "attributes": {
          "properties": {
            "username": {
              "type": "string"
            },
            "first_name": {
              "type": "string"
            },
            "last_name": {
              "type": "string"
            },
            "groups": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "api_token": {
              "type": "string",
              "readOnly": true
            },
            "api_key_expires_at": {
              "type": "string",
              "format": "date-time",
              "readOnly": true
            },
            "admin": {
              "type": "boolean",
              "readOnly": true
            }
          }
        },
        "links": {
          "properties": {
            "self": {
              "type": "string",
              "format": "url",
              "readOnly": true
            }
          }
        }
      }
    },
...
  },
  "host": "localhost:3000"
}